% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_flags.hybcap}
\alias{add_flags.hybcap}
\title{Hybrid capture method function}
\usage{
\method{add_flags}{hybcap}(
  mdlvalr_list,
  pipeline = "hybcap",
  cds_table = NULL,
  var_pass_fail_logic = "TRUE",
  cov_pass_fail_logic = "TRUE"
)
}
\arguments{
\item{mdlvalr_list}{Name of the list to flag input data}

\item{pipeline}{\link{string} Name of pipeline that data were derived from.}
}
\value{
Returns a named list of tibbles
}
\description{
Hybrid capture method function
}
