% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_flags}
\alias{add_flags}
\alias{add_flags.hybcap}
\title{Add new flag columns to variant and coverage tables}
\usage{
add_flags(
  mdlvalr_list,
  pipeline = NULL,
  cds_table = NULL,
  var_pass_fail_logic = "TRUE",
  cov_pass_fail_logic = "TRUE"
)

\method{add_flags}{hybcap}(
  mdlvalr_list,
  pipeline = "hybcap",
  cds_table = NULL,
  var_pass_fail_logic = "TRUE",
  cov_pass_fail_logic = "TRUE"
)
}
\arguments{
\item{mdlvalr_list}{A standard \code{mdlvalr_list} object (as generated by \code{get_stats()}).
The class of this object can be used to automatically determine which
function methods will be used.}

\item{pipeline}{Standard name of the pipeline from which the data were derived.}

\item{cds_table}{A tibble object that contains the RefSeq CDS info. See also \code{\link[=add_cds_column]{add_cds_column()}}.}

\item{var_pass_fail_logic}{A logic statement that can be applied to the variant input table, such that TRUE returns a "pass" and FALSE returns a "fail" character string. Default is "TRUE", resulting in all variants getting a "pass" label.}

\item{cov_pass_fail_logic}{A logic statement that can be applied to the coverage input table, such that TRUE returns a "pass" and FALSE returns a "fail" character string. Default is "TRUE", resulting in all variants getting a "pass" label.}
}
\value{
Returns a standard \code{mdlvalr_list} object (i.e. a regular R list of lists) that contains tables as named list elements.
The class of the retuned object is updated to include the name of the pipeline. This facilitates use of alternative methods in
downstream functions.
}
\description{
Add new flag columns to variant and coverage tables
}
\section{Methods}{

This function is a \strong{generic}, which means that different
implementations (methods) can be executed for different classes. The class
of the first argument will determine which method is used. See the documentation of
individual methods for extra arguments and differences in behaviour.
}

\examples{
\dontrun{
class(mdlvalr_list)
# [1] "hybcap" "list"
mdlvalr_list <- add_flags(mdlvalr_list,
    cds_table,
    var_pass_fail_logic = "VAF >= 0.05",
    cov_pass_fail_logic = "fraction_125x >= 0.9")
}

}
