% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var_labels.R
\name{add_var_labels.hybcap}
\alias{add_var_labels.hybcap}
\title{Hybrid capture method function}
\usage{
\method{add_var_labels}{hybcap}(mdlvalr_list, pipeline = "hybcap")
}
\arguments{
\item{mdlvalr_list}{Name of the list to flag input data}

\item{pipeline}{\link{string} Name of pipeline that data were derived from.}
}
\value{
Returns a named list of tibbles
}
\description{
Hybrid capture method function
}
