% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{all_sample_compare}
\alias{all_sample_compare}
\title{Compare a set of samples for concordance}
\usage{
all_sample_compare(
  samples_tbl,
  sample_group,
  comparison_group,
  var_path,
  cov_path,
  min_vaf = 0.05,
  min_fraction_125x = 0.9
)
}
\arguments{
\item{samples_tbl}{Tibble of sample info.}

\item{sample_group}{String of column name in \code{samples_tbl} that represents the sample group for comparison. This column in samples_tbl needs to be a factor.}

\item{comparison_group}{String of column name in \code{samples_tbl} that represents the comparison group. This column in samples_tbl needs to be a factor with exactly two levels. For example, levels might be c("ver1", "ver2") or c("20ng", "50ng").}

\item{var_path}{String of column name in \code{samples_tbl} that represents the variant file path.}

\item{cov_path}{String of column name in \code{samples_tbl} that represents the coverage file path.}

\item{min_vaf}{Variant tables are filtered based on minimum variant allele frequency (VAF). Variants with VAF less than or equal to this value are not used in comparison.  \link{default: 0.05}.}

\item{min_fraction_125x}{Exon tables are filtered based on the minimum fraction of bases covered at 125x. \link{default: 0.9}.}
}
\value{
Returns a named list of lists. The length is equal to the number of comparisons made.
}
\description{
This function takes a tibble of sample info (filename paths, sample names, etc.) and runs a concordance analysis.
}
