% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vars.R
\name{compare_vars.hybcap}
\alias{compare_vars.hybcap}
\title{Compare two variant call sets}
\usage{
\method{compare_vars}{hybcap}(
  mdlvalr_list,
  pipeline = "hybcap",
  match_by_colnames = c("CHROM", "POS", "REF", "ALT", "GENE", "EXON"),
  suffix_names = c(".sample1", ".sample2")
)
}
\arguments{
\item{match_by_colnames}{Vector of column names to use for matching. Passed to \verb{by = } parameter in join functions. \link{default: c("CHROM", "POS", "REF", "ALT", "GENE", "EXON")}.}

\item{suffix_names}{Suffix names added to column names after running join functions. Vector of length two, \link{default: c(".sample_1", ".sample_2")}.}

\item{sample_1_name}{String for sample_1 name.}

\item{sample_1_var}{Tibble of sample_1 variants.}

\item{sample_1_cov}{Tibble of sample_1 exon coverage.}

\item{sample_2_name}{String for sample_2 name.}

\item{sample_2_var}{Tibble of sample_2 variants.}

\item{sample_2_cov}{Tibble of sample_2 exon coverage.}

\item{min_vaf}{Variant tables are filtered based on minimum variant allele frequency (VAF). Variants with VAF less than or equal to this value are not used in comparison.  \link{default: 0.05}.}

\item{min_fraction_125x}{Exon tables are filtered based on the minimum fraction of bases covered at 125x. \link{default: 0.9}.}
}
\value{
Returns a named list of tibbles or integers tibble sizes.
}
\description{
This function takes two samples as input and compares them using \verb{dplyr::*_join()} functions. The results of the joins are captured and returned in a list. This is an internal function.
}
