% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel.hybcap}
\alias{export_excel.hybcap}
\title{Export data to Excel files using tabs}
\usage{
\method{export_excel}{hybcap}(mdlvalr_list, pipeline = "hybcap", filename_prefix = "")
}
\description{
Export data to Excel files using tabs
}
