% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Generic function}
\usage{
export_excel(mdlvalr_list, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
