% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data.hybcap}
\alias{filter_data.hybcap}
\title{Hybrid capture method function}
\usage{
\method{filter_data}{hybcap}(mdlvalr_list, pipeline = "hybcap")
}
\arguments{
\item{mdlvalr_list}{Name of the list to flag input data}

\item{pipeline}{\link{string} Name of pipeline that data were derived from.}
}
\value{
Returns a named list of tibbles
}
\description{
Hybrid capture method function
}
