% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{fix_names}
\alias{fix_names}
\title{Fix a vector of names to fit into Excel tabs}
\usage{
fix_names(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{Vector of string names.}
}
\value{
Returns a vector of shortened strings.
}
\description{
This function takes a vector of string names and reduces their length to be less than 31 characters (the max for Excel tabs).
}
