% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data.hybcap}
\alias{get_data.hybcap}
\title{Hybrid capture method function}
\usage{
\method{get_data}{hybcap}(sample_sheet, pipeline = "hybcap")
}
\arguments{
\item{sample_sheet}{Name of the tibble or data.frame that contains sample info.}

\item{pipeline}{\link{string} Name of pipeline that data were derived from.}
}
\value{
Returns a named list of tibbles (length = 2). The first is the variant table and the second is the exon coverage table.
}
\description{
Hybrid capture method function
}
