% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats.hybcap}
\alias{get_stats.hybcap}
\title{Capture summary stats from the comparisons}
\usage{
\method{get_stats}{hybcap}(mdlvalr_list, pipeline = "hybcap")
}
\arguments{
\item{flag_column_pass}{Column name to be used for filtering variant calls that pass or not. Only variants rows with a "yes" in this column will be used for calculating pass variants. Another possible option is "vaf_gt_0.05".}
}
\value{
Returns a named list of tibbles or integers tibble sizes.
}
\description{
Capture summary stats from the comparisons
}
