% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{out_files}
\alias{out_files}
\title{Create output files after concordance was performed}
\usage{
out_files(two_sample_compare_list, filename_prefix = "")
}
\arguments{
\item{two_sample_compare_list}{List of data generated by \code{two_sample_compare()}.}

\item{filename_prefix}{String or path that is prepended to exported filenames. \link{default = ""}.}
}
\value{
Returns a vector of shortened strings.
}
\description{
This function takes the output list from \code{two_sample_compare()} as input and writes output Excel files.
}
