% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{read_files}
\alias{read_files}
\title{Read Excel files into R tibbles}
\usage{
read_files(var_file, cov_file)
}
\arguments{
\item{var_file}{File path to the variant file. File path supplied to \code{openxlsx::read.xlsx()}.}

\item{cov_file}{File path to the exon coverage file, sheet = 2. File path supplied to \code{openxlsx::read.xlsx()}.}
}
\value{
Returns a named list of tibbles (length = 2). The first is the variant table and the second is the exon coverage table.
}
\description{
The function reads the variant list and exon coverage Excel files and returns tibbles. The is an internal function used by \code{all_sample_compare()}.
}
