% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{two_sample_compare}
\alias{two_sample_compare}
\title{Compare two variant call sets}
\usage{
two_sample_compare(
  sample_1_name,
  sample_1_var,
  sample_1_cov,
  sample_2_name,
  sample_2_var,
  sample_2_cov,
  min_vaf = 0.5,
  min_exon_coverage = 125,
  match_by_colnames = c("CHROM", "POS", "REF", "ALT", "GENE", "EXON"),
  suffix_names = c(".sample_1", ".sample_2")
)
}
\arguments{
\item{sample_1_name}{String for sample_1 name.}

\item{sample_1_var}{Tibble of sample_1 variants.}

\item{sample_1_cov}{Tibble of sample_1 exon coverage.}

\item{sample_2_name}{String for sample_2 name.}

\item{sample_2_var}{Tibble of sample_2 variants.}

\item{sample_2_cov}{Tibble of sample_2 exon coverage.}

\item{min_vaf}{Variant tables are filtered based on minimum variant allele frequency (VAF). Variants with VAF less than or equal to this value are not used in comparison.  \link{default: 0.05}.}

\item{min_exon_coverage}{Variant tables are filtered based on minimum exon coverage level. Variants with coverage less than or equal to this value are not used in comparison. \link{default: 125}.}

\item{match_by_colnames}{Vector of column names to use for matching. Passed to \verb{by = } parameter in join functions. \link{default: c("CHROM", "POS", "REF", "ALT", "GENE", "EXON")}.}

\item{suffix_names}{Suffix names added to column names after running join functions. Vector of length two, \link{default: c(".sample_1", ".sample_2")}.}
}
\value{
Returns a named list of tibbles or integers tibble sizes.
}
\description{
This function takes two samples as input and compares them using \verb{dplyr::*_join()} functions. The results of the joins are captured and returned in a list. This is an internal function.
}
