% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_samples.R
\name{validate_samples}
\alias{validate_samples}
\title{Validate a table of sample info @param samples_tbl Tibble of sample info. Requires five columns (listed below).}
\usage{
validate_samples(
  samples_tbl,
  sample_name,
  sample_group,
  comparison_group,
  var_path,
  cov_path,
  pipeline
)
}
\arguments{
\item{sample_name}{Column name that specifies the sample names. \link{string}.}

\item{sample_group}{Column name that specifies the sample_group. \link{string}.}

\item{comparison_group}{Column name that specifies the comparison_group. \link{string}.}

\item{var_path}{Column name that specifies the variants table file path. \link{string}.}

\item{cov_path}{Column name that specifies the coverage table file path. \link{string}.}

\item{pipeline}{Name of pipeline that produced the data. One of c("hybcap", "exome", "germline").}
}
\description{
Validate a table of sample info @param samples_tbl Tibble of sample info. Requires five columns (listed below).
}
\examples{
invisible(samples_tbl)
sample_sheet <- validate_samples(samples_tbl, sample_name = "sample_name", sample_group = "sample_name_short", comparison_group = "group", var_path = "filtered_path", cov_path = "coverage_path", pipeline = "hybcap")

}
