% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_cds_column}
\alias{add_cds_column}
\title{Add a column indicating whether variant overlaps with a CDS region}
\usage{
add_cds_column(var_tbl, cds_table)
}
\arguments{
\item{var_tbl}{Tibble of variants. Must have columns named: chr, start, end, type.}

\item{cds_table}{Tibble RefSeq gff that includes CDS regions. Must have column name: "type" that indicates the CDS or not.}
}
\value{
Returns the same var_tbl that was input, but includes additional column for CDS.
}
\description{
Add a column indicating whether variant overlaps with a CDS region
}
