% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_flags}
\alias{add_flags}
\title{Generic function}
\usage{
add_flags(mdlvalr_list, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
