% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var_labels.R
\name{add_var_labels}
\alias{add_var_labels}
\title{Generic function}
\usage{
add_var_labels(mdlvalr_list, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
