% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vars.R
\name{compare_vars}
\alias{compare_vars}
\title{Generic function}
\usage{
compare_vars(mdlvalr_list, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
