% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\alias{export_excel.hybcap}
\title{Write Excel files containing all sumarized data within tabs}
\usage{
export_excel(mdlvalr_list, pipeline = NULL, filename_prefix = "")

\method{export_excel}{hybcap}(mdlvalr_list, pipeline = "hybcap", filename_prefix = "")
}
\arguments{
\item{mdlvalr_list}{A standard \code{mdlvalr_list} object (as generated by \code{compare_data()}).
The class of this object can be used to automatically determine which
function methods will be used.}

\item{pipeline}{Standard name of the pipeline from which the data were derived.}

\item{filename_prefix}{String or filepath that will be glued as a prefix to the output filename.}
}
\value{
Returns a standard \code{mdlvalr_list} object (i.e. a regular R list of lists) that contains tables as named list elements.
The class of the retuned object is updated to include the name of the pipeline. This facilitates use of alternative methods in
downstream functions.
}
\description{
Write Excel files containing all sumarized data within tabs

Export data to Excel files using tabs
}
\section{Methods}{

This function is a \strong{generic}, which means that different
implementations (methods) can be executed for different classes. The class
of the first argument will determine which method is used. See the documentation of
individual methods for extra arguments and differences in behaviour.
}

\examples{
class(mdlvalr_list)
# [1] "hybcap" "list"
\notrun{
mdlvalr_list <- export_excel(
    mdlvalr_list,
    pipeline = "hybcap",
    filename_prefix = "logic_VAF_gt_0.05_")
}
}
