% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Generic function}
\usage{
get_data(sample_sheet, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
