% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read pipeline files into a list}
\usage{
get_data(sample_sheet, pipeline = c("hybcap"))
}
\arguments{
\item{sample_sheet}{Name of the tibble or data.frame that contains sample
info. One set of comparisons per row of table. There are five required
column names: (1) \code{comparison_name} (e.g. "sample1_run1_vs_sample1_run2").
(2, 3) \code{var_path_1} and \code{var_path_2} that provide complete paths to the
variant files. (4, 5) \code{cov_path_1} and \code{cov_path_2} that provide complete
paths to the coverage files.}

\item{pipeline}{Standard name of the pipeline from which the data were derived}
}
\value{
Returns a standard \code{mdlvalr_list} object (i.e. a regular R list of lists) that contains tables as named list elements.
The class of the retuned object is updated to include the name of the pipeline. This facilitates use of alternative methods in
downstream functions.
}
\description{
Read pipeline files into a list
}
\examples{
\dontrun{
mdlvalr_list <- get_data(sample_sheet, pipeline = "hybcap")
}
}
