% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Generic function}
\usage{
get_stats(mdlvalr_list, ...)
}
\description{
This generic function can take only the first argument as input and use the correct "method" automatically based on the class of the argument.
}
